[MIG] website_form_require_legal: Migration to 17.0
