+++ website_form_require_legal/README.rst
@@ -82,6 +82,10 @@ Contributors
 
    -  Carlos Roca
 
+-  `NICO SOLUTIONS <https://www.nico-solutions.de>`__:
+
+   -  Nils Coenen
+
 Maintainers
 -----------
