+++ website_form_require_legal/readme/CONTRIBUTORS.md
@@ -1,2 +1,4 @@
 - [Tecnativa](https://www.tecnativa.com):
   - Carlos Roca
+- [NICO SOLUTIONS](https://www.nico-solutions.de):
+  - Nils Coenen
